Refactor tr_msg and tr_filter to disentangle protocols
Dependencies between the tr_msg and tr_filter modules were
preventing building a libtr_tid.la library that did not include
far more of the trust router than it needed. It also prevented
using that library to provide service to the trust router
without including redundant static copies in the trust_router
binary.

  * Separate protocols out of tr_msg.c
    - replace switch() blocks with a dynamic handler table
    - move protocol-specific code into mon/trp/tid directories
  * Separate trp fields out of tr_filter.c
    - replace static handler table with a dynamic table
    - move trp fields to trp/trp_filter.c
    - leave tid in tr_filter.c for now, we don't currently have
      any use for a TID-less filter module
  * Rearrange Makefile.am
