+++ include/tr_filter.h
@@ -58,10 +58,10 @@ typedef enum tr_filter_action {
 
 /* Filter types */
 typedef enum {
-  TR_FILTER_TYPE_TID_INBOUND = 0,
+  TR_FILTER_TYPE_UNKNOWN = 0, /* make this 0 so zeroed fields are TYPE_UNKNOWN */
+  TR_FILTER_TYPE_TID_INBOUND,
   TR_FILTER_TYPE_TRP_INBOUND,
   TR_FILTER_TYPE_TRP_OUTBOUND,
-  TR_FILTER_TYPE_UNKNOWN
 } TR_FILTER_TYPE;
 
 typedef struct tr_fspec {
@@ -145,13 +145,20 @@ typedef TR_LIST_ITER TR_FLINE_ITER;
 #define tr_fline_iter_next(ITER) (tr_list_iter_next(ITER))
 #define tr_fline_add_spec(LINE, SPEC) ((TR_NAME *) tr_list_add((LINE)->specs, (SPEC), 1))
 
+/* CMP functions return values like strcmp: 0 on match, <0 on target<val, >0 on target>val */
+typedef int (TR_FILTER_FIELD_CMP)(TR_FILTER_TARGET *target, TR_NAME *val);
+/* get functions return TR_NAME format of the field value. Caller must free it. */
+typedef TR_NAME *(TR_FILTER_FIELD_GET)(TR_FILTER_TARGET *target);
+
+
 /*In tr_constraint.c and exported, but not really a public symbol; needed by tr_filter.c and by tr_constraint.c*/
 int TR_EXPORT tr_prefix_wildcard_match(const char *str, const char *wc_str);
 
 int tr_filter_apply(TR_FILTER_TARGET *target, TR_FILTER *filt, TR_CONSTRAINT_SET **constraints, TR_FILTER_ACTION *out_action);
-void tr_filter_target_free(TR_FILTER_TARGET *target);
+
+TR_FILTER_TARGET TR_EXPORT *tr_filter_target_new(TALLOC_CTX *mem_ctx);
+void TR_EXPORT tr_filter_target_free(TR_FILTER_TARGET *target);
 TR_FILTER_TARGET *tr_filter_target_tid_req(TALLOC_CTX *mem_ctx, TID_REQ *req);
-TR_FILTER_TARGET *tr_filter_target_trp_inforec(TALLOC_CTX *mem_ctx, TRP_UPD *upd, TRP_INFOREC *inforec);
 
 TR_CONSTRAINT_SET *tr_constraint_set_from_fline(TR_FLINE *fline);
 
@@ -159,6 +166,11 @@ int tr_filter_validate(TR_FILTER *filt);
 int tr_filter_validate_spec_field(TR_FILTER_TYPE ftype, TR_FSPEC *fspec);
 const char *tr_filter_type_to_string(TR_FILTER_TYPE ftype);
 TR_FILTER_TYPE tr_filter_type_from_string(const char *s);
+
+int TR_EXPORT tr_filter_add_field_handler(TR_FILTER_TYPE ftype,
+                                          const char *name,
+                                          TR_FILTER_FIELD_CMP *cmp,
+                                          TR_FILTER_FIELD_GET *get);
 
 /* tr_filter_encoders.c */
 json_t *tr_filter_set_to_json(TR_FILTER_SET *filter_set);
