+++ mon/monc.c
@@ -54,6 +54,8 @@ MONC_INSTANCE *monc_new(TALLOC_CTX *mem_ctx)
     }
 
     monc->gssc->service_name = "trustmonitor";
+
+    mon_tr_msg_init(); /* Prepare to send messages */
   }
   return monc;
 }
@@ -81,7 +83,6 @@ MON_RESP *monc_send_request(TALLOC_CTX *mem_ctx, MONC_INSTANCE *monc, MON_REQ *r
   if (!(msg = talloc_zero(tmp_ctx, TR_MSG)))
     goto cleanup;
 
-  msg->msg_type = MON_REQUEST;
   tr_msg_set_mon_req(msg, req);
 
   resp_msg = tr_gssc_exchange_msgs(tmp_ctx, monc->gssc, msg);
