+++ trust_router.spec
@@ -1,7 +1,7 @@
 %global optflags %{optflags} -Wno-parentheses
 %{!?_release_number: %define _release_number 1}
 Name:           trust_router
-Version:        3.4.1~2
+Version:        3.4.1~3
 Release:        %{_release_number}%{?dist}
 Summary:        Moonshot Trust Router
